Split image metadata facade.
