+++ api/facadeversions.go
@@ -38,7 +38,8 @@ var facadeVersions = map[string]int{
 	"HighAvailability":             2,
 	"HostKeyReporter":              1,
 	"ImageManager":                 2,
-	"ImageMetadata":                2,
+	"ImageMetadata":                3,
+	"ImageMetadataManager":         1,
 	"InstancePoller":               3,
 	"KeyManager":                   1,
 	"KeyUpdater":                   1,
