+++ apiserver/facades/agent/imagemetadata/functions_test.go
@@ -9,7 +9,7 @@ import (
 	jc "github.com/juju/testing/checkers"
 	gc "gopkg.in/check.v1"
 
-	"github.com/juju/juju/apiserver/facades/client/imagemetadata"
+	"github.com/juju/juju/apiserver/facades/agent/imagemetadata"
 	"github.com/juju/juju/apiserver/params"
 	"github.com/juju/juju/testing"
 )
