+++ apiserver/facades/client/imagemetadatamanager/export_test.go
@@ -1,0 +1,8 @@
+// Copyright 2015 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package imagemetadatamanager
+
+var (
+	CreateAPI = createAPI
+)
